ci: Add linter CI and lint all codes (#156)
* ci: add linter


* Update and rename main.yml to linter.yml


* Update linter.yml


* Update linter.yml


* Update linter.yml


* Update linter.yml


* Update linter.yml


* cargo fmt

* Update linter.yml


* Update linter.yml


* add cargo fmt ignore

* fix: cargo fmt ignore by using rust nightly

* fix

* revert to stable toolchain

* fix

* fix clippy

* add: Lint script in justfile

---------

+++ database/src/actions/analysis.rs
@@ -420,7 +420,7 @@ pub async fn get_percentile(
     };
     // .with_context(|| "Unable to get analysis value")?;
 
-    Ok(result.unwrap_or_default() as f32)
+    Ok(result.unwrap_or_default())
 }
 
 pub async fn get_percentile_analysis_result(
